Phase 2: x86_64 boot (multiboot2 + EFI GRUB) through the full generic init
Add arch/x86_64: higher-half linker script, multiboot2 header + 32-bit
long-mode trampoline (identity + direct-map page tables, physical->virtual
entry stub), 64-bit GDT/TSS/IDT + syscall/sysret MSRs, 4-level paging +
arch_mm, multiboot2 memory-map detection, PIT timer, APIC/PIC IRQ stack,
uaccess, exception handler.  The shared kernel init now runs under
qemu-system-x86_64 + OVMF + EFI GRUB, mounts ext2 and spawns init
(64-bit userspace is phase 3).

Also: vmalloc.c address types fully widened (64-bit ioremap), slab.c
pointer truncation fixed, mminit log strings use arch constants.
i386 regression boot verified.

+++ kernel/process.c
@@ -807,7 +807,7 @@ static int do_exec(const char *user_path, char **user_argv, char **user_envp)
     }
     
     /* Build the argc/argv/envp block on the new stack and jump to userspace */
-    uint32_t user_esp;
+    uintptr_t user_esp;
     if (setup_initial_stack(stack_pages, stack_vaddr, &argv, &envp,
                             &user_esp) < 0) {
         printk("[EXEC] Initial stack too small for argv/envp\n");
